Add new feature groups strategy
This commit adds an additional feature group cross validation strategy.
Additional to the leave_one_in and leave_one_out strategies, we add a
strategy that includes all possible combinations (with replacement) of
the feature groups. This can allow us to compare the behavior of models
with different feature spaces, and the marginal gain/leverage of each
feature groups ceteris paribus.